requirements.txt
langchain==0.0.340
pandas==2.1.3
Pillow==10.1.0
PyPDF2==3.0.1
python_docx==1.1.0
streamlit==1.28.2
streamlit_chat==0.1.1
langchain_experimental==0.0.41
openai==1.35.10
SQLAlchemy==1.4.48
sqlalchemy-bigquery
db-dtypes
google-api-core==2.14.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-speech==2.22.0
tabulate